Move _CoordinateIndexer to base module
With both geoseries and geodataframe modules using the class, base.py is
a better place to define the _CoordinateIndexer class.

Also add docs to the _CoordinateIndexer class.
